frontend: Consolidate remaining registry packages
This PR simplifies and consolidates the remaining parts of the various registry packages, and moves everything under cmd/frontend/internal for better encapusation.
This also removes the need to _ import packages which feels very brittle.

Test plan: Go compiler doesn't complain about moved code.
