+++ cmd/frontend/internal/httpapi/BUILD.bazel
@@ -33,10 +33,10 @@ go_library(
         "//cmd/frontend/internal/httpapi/releasecache",
         "//cmd/frontend/internal/httpapi/webhookhandlers",
         "//cmd/frontend/internal/modelconfig",
+        "//cmd/frontend/internal/registry",
         "//cmd/frontend/internal/routevar",
         "//cmd/frontend/internal/search",
         "//cmd/frontend/internal/ssc",
-        "//cmd/frontend/registry/api",
         "//cmd/frontend/webhooks",
         "//internal/actor",
         "//internal/api",
