+++ cmd/frontend/internal/licensing/init/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "//cmd/frontend/internal/dotcom/productsubscription",
         "//cmd/frontend/internal/licensing/enforcement",
         "//cmd/frontend/internal/licensing/resolvers",
-        "//cmd/frontend/internal/registry",
         "//internal/codeintel",
         "//internal/conf/conftypes",
         "//internal/database",
