+++ cmd/frontend/internal/registry/http_api_test.go
@@ -18,7 +18,7 @@ func TestHandleRegistry(t *testing.T) {
 		rr.Body = new(bytes.Buffer)
 		req, _ := http.NewRequest("GET", "/.api/registry/extensions", nil)
 		req.Header.Set("Accept", "application/vnd.sourcegraph.api+json;version=20180621")
-		handleRegistry(rr, req)
+		HandleRegistry(rr, req)
 		if want := 200; rr.Result().StatusCode != want {
 			t.Errorf("got HTTP status %d, want %d", rr.Result().StatusCode, want)
 		}
@@ -33,7 +33,7 @@ func TestHandleRegistry(t *testing.T) {
 		rr.Body = new(bytes.Buffer)
 		req, _ := http.NewRequest("GET", "/.api/registry/extensions/extension-id/sourcegraph/go", nil)
 		req.Header.Set("Accept", "application/vnd.sourcegraph.api+json;version=20180621")
-		handleRegistry(rr, req)
+		HandleRegistry(rr, req)
 		if want := 200; rr.Result().StatusCode != want {
 			t.Errorf("got HTTP status %d, want %d", rr.Result().StatusCode, want)
 		}
