+++ cmd/frontend/internal/registry/types.go
@@ -1,4 +1,4 @@
-package client
+package registry
 
 import "time"
