--- cmd/frontend/registry/api/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "api",
-    srcs = [
-        "doc.go",
-        "extensions_helpers.go",
-        "gating.go",
-        "http_api.go",
-    ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/registry/api",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//cmd/frontend/registry/client",
-        "//internal/dotcom",
-        "//lib/errors",
-    ],
-)
