--- cmd/frontend/registry/api/doc.go
@@ -1,2 +1,0 @@
-// Package api contains a partial implementation of the extension registry.
-package api
