--- cmd/frontend/registry/client/BUILD.bazel
@@ -1,19 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "client",
-    srcs = [
-        "client.go",
-        "doc.go",
-        "types.go",
-    ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/registry/client",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//internal/httpcli",
-        "//internal/trace",
-        "//lib/errors",
-        "@com_github_google_uuid//:uuid",
-        "@io_opentelemetry_go_otel//attribute",
-    ],
-)
