--- cmd/frontend/registry/client/doc.go
@@ -1,2 +1,0 @@
-// Package client contains a partial implementation of the extension registry.
-package client
