+++ cmd/frontend/shared/BUILD.bazel
@@ -33,14 +33,12 @@ go_library(
         "//cmd/frontend/internal/own",
         "//cmd/frontend/internal/prompts",
         "//cmd/frontend/internal/rbac",
-        "//cmd/frontend/internal/registry",
         "//cmd/frontend/internal/repos/webhooks",
         "//cmd/frontend/internal/savedsearches",
         "//cmd/frontend/internal/scim",
         "//cmd/frontend/internal/search",
         "//cmd/frontend/internal/searchcontexts",
         "//cmd/frontend/internal/telemetry",
-        "//cmd/frontend/registry/api",
         "//internal/codeintel",
         "//internal/codeintel/shared",
         "//internal/conf",
